Add extend core (#816)
* add exit?

* use working branch

* use core 2.0.0

* update changelog

+++ android_tests/lib/android/specs/common/command/command.rb
@@ -1,7 +1,7 @@
 # rake android[common/command/command]
 describe 'common/command/command' do
   t 'command' do
-    File.delete 'logcat.log'
+    File.delete 'logcat.log' if File.exist? 'logcat.log'
 
     start_logs_broadcast
     sleep 5
